requirements.txt
streamlit
requests
pandas
python-dotenv
openai